// src/lib/admin-section-schema.ts
export type AdminFieldType =
  | 'text'
  | 'textarea'
  | 'richtext'
  | 'number'
  | 'boolean'
  | 'select'
  | 'date'
  | 'color'
  | 'image'
  | 'url'
  | 'array'
  | 'object'